+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.2
 - language server support for nested Alan project directories: the extension starts a language server instance for each `versions.json` and `project.json` that is found
+- only show (applicable) snippets in `application.alan` and `interface.alan` files
+- snippets change: exclude `: `
 
 ## 1.0.1
 - language server support for Windows